Separate out mask helper functions
