+++ packages/design-system/src/components/TextField/index.js
@@ -1,1 +1,2 @@
-export { default as TextField, unmaskValue } from './TextField';
+export { default as TextField } from './TextField';
+export { unmaskValue } from './maskHelpers';
